+++ NEWS.md
@@ -21,4 +21,4 @@
 # 2019-07-24 - popkinsuppl 1.0.4.9000
 
 * Adapted package to use new `solve_m_mem_lim` from the `popkin` package, which generalizes some of the calculations that used to be part of the memory estimation/control code in `get_mem_lim_m_kinship_std` and `get_mem_lim_m_WC`.
-
+* `kinship_std` now preserves individual names when present in input genotype matrix.